Rename network_helper -> network
