+++ tests/test_discovery.py
@@ -5,7 +5,7 @@ from greeclimate.device import DeviceInfo
 from unittest.mock import patch
 
 @pytest.mark.asyncio
-@patch("greeclimate.network_helper.search_devices")
+@patch("greeclimate.network.search_devices")
 async def test_discover_device(mock_search_devices):
     mock_search_devices.return_value = [
         ("1.1.1.0", "7000", "aabbcc001122", "MockDevice1")
@@ -17,7 +17,7 @@ async def test_discover_device(mock_search_devices):
     assert len(devices) > 0
 
 @pytest.mark.asyncio
-@patch("greeclimate.network_helper.search_devices")
+@patch("greeclimate.network.search_devices")
 async def test_discover_no_devices(mock_search_devices):
     mock_search_devices.return_value = []
